// src/payload-types.ts
export interface Media {
  id: number;
  /**
   * Lue par les lecteurs d’écran et utile au référencement.
   */
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    vignette?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    carte?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Evenement {
  id: number;
  nom: string;
  dateDebut: string;
  dateFin?: string | null;
  enLigne?: boolean | null;
  lieu?: string | null;
  lienVisio?: string | null;
  /**
   * Affichée sur la carte de l’événement (280 caractères maximum).
   */
  description: string;
  prix: 'gratuit' | 'payant';
  montant?: string | null;
  public: 'tout_public' | 'ipestiens_alumni' | 'ipestiens' | 'alumni';
  lienInscription?: string | null;
  photo?: (number | null) | Media;
  contenu?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  /**
   * Photos affichées sur la page détaillée, par exemple après l’événement.
   */
  galerie?: (number | Media)[] | null;
  finEffective?: string | null;
  updatedAt: string;
  createdAt: string;
}